requirements/base.txt
pyyaml
matplotlib==3.6.3
numpy==1.24.1
opencv_python==4.7.0.72
tensorflow==2.11.0
